sub 2 - fixing bug

+++ open-music-api/src/api/playlists/handler.js
@@ -39,6 +39,8 @@ class PlaylistsHandler {
 
   async deletePlaylistByIdHandler (request, h) {
     const { id } = request.params
+    const { id: credentialId } = request.auth.credentials
+    await this._service.verifyPlaylistOwner(id, credentialId)
     await this._service.deletePlaylistById(id)
     return {
       status: 'success',
